requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
python-dotenv>=0.19.0
openai>=1.0.0
perplexity>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
habanero>=1.2.3
arxiv>=2.0.0 